jobseeker-backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
contextlib2==21.6.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-cors-headers==4.1.0
django-ninja==0.22.2
django-ninja-extra==0.19.1
django-ninja-jwt==5.2.5
httpie==3.2.2
idna==3.4
injector==0.21.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
Pillow==10.0.0
pycparser==2.21
pydantic==1.10.12
Pygments==2.15.1
PyJWT==2.8.0
PySocks==1.7.1
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
